services: iiif: refactor; add methods to storage
