+++ invenio_rdm_records/services/iiif/tasks.py
@@ -12,7 +12,7 @@ from invenio_db import db
 
 from invenio_rdm_records.proxies import current_rdm_records_service
 
-from .utils import tiles_storage as tif_store # TODO Change to singleton
+from .storage import tiles_storage as tif_store # TODO Change to singleton
 
 @shared_task(
     ignore_result=True,
@@ -21,12 +21,15 @@ def generate_tiles(record_id, file_key, params=None):
     """Generate pyramidal tiff."""
     record = current_rdm_records_service.record_cls.pid.resolve(record_id)
     status_file = record.media_files[file_key + ".ptif"]
-    status_file["processor"]["status"] = "processing"
+    status_file.processor["status"] = "processing"
     status_file.commit()
     db.session.commit()
 
     conversion_state = tif_store.save(record, file_key)
 
-    status_file["processor"]["status"] = "finished" if conversion_state else "failed"
+    status_file.processor["status"] = "finished" if conversion_state else "failed"
+    status_file.file.file_model.uri = str(
+        tif_store._get_file_path(record, file_key)
+    )
     status_file.commit()
     db.session.commit()
